Refactor memory utilities and update MaskBuffer logic
Moved AlignedMemoryResource and MemoryBlockPool to a new Memory namespace and directory, updating all references accordingly. Introduced a Memory interface library in CMake and linked it to SelfieSegmenter. Removed ForceAlignmentResource and refactored MaskBuffer to use the new AlignedMemoryResource, expanding its buffer logic to use three buffers and atomic indices for improved concurrency.
